Refactor Actor interface to abstract class
- Convert Actor interface to an abstract class.
- Centralize shared fields (uuid, name, health, dead, tile, status, etc.) in Actor.
- Update CardPlayer and Structure hierarchies to use the new base class.
- Remove redundant field and method overrides across actor implementations.
