+++ nomad-realms/src/main/java/nomadrealms/context/game/actor/types/cardplayer/Farmer.java
@@ -28,8 +28,6 @@ import nomadrealms.render.RenderingEnvironment;
 
 public class Farmer extends CardPlayer {
 
-	private final String name;
-
 	/**
 	 * No-arg constructor for serialization.
 	 */
@@ -108,11 +106,6 @@ public class Farmer extends CardPlayer {
 	@Override
 	public List<Appendage> appendages() {
 		return asList(HEAD, EYE, EYE, TORSO, ARM, ARM, LEG, LEG);
-	}
-
-	@Override
-	public String name() {
-		return name;
 	}
 
 }
