+++ nomad-realms/src/main/java/nomadrealms/context/game/actor/types/cardplayer/FeralMonkey.java
@@ -26,8 +26,6 @@ import nomadrealms.render.RenderingEnvironment;
 
 public class FeralMonkey extends CardPlayer {
 
-	private final String name;
-
 	/**
 	 * No-arg constructor for serialization.
 	 */
@@ -80,11 +78,6 @@ public class FeralMonkey extends CardPlayer {
 	@Override
 	public List<Appendage> appendages() {
 		return asList(HEAD, EYE, EYE, TORSO, ARM, ARM, LEG, LEG, TAIL);
-	}
-
-	@Override
-	public String name() {
-		return name;
 	}
 
 }
