+++ nomad-realms/src/main/java/nomadrealms/context/game/actor/types/cardplayer/VillageChief.java
@@ -23,8 +23,6 @@ import nomadrealms.render.RenderingEnvironment;
 
 public class VillageChief extends CardPlayer {
 
-	private final String name;
-
 	public VillageChief(String name) {
 		this.setAi(new VillageChiefAI(this));
 		this.name = name;
